authz: add test to prevent new endpoints not covered by authz (#785)
