+++ nexus/Cargo.toml
@@ -120,6 +120,7 @@ nexus-test-utils-macros = { path = "test-utils-macros" }
 nexus-test-utils = { path = "test-utils" }
 omicron-test-utils = { path = "../test-utils" }
 openapiv3 = "1.0"
+regex = "1.5.5"
 subprocess = "0.2.8"
 
 [dev-dependencies.openapi-lint]
