+++ nexus/tests/integration_tests/mod.rs
@@ -18,6 +18,7 @@ mod router_routes;
 mod subnet_allocation;
 mod timeseries;
 mod unauthorized;
+mod unauthorized_coverage;
 mod updates;
 mod users_builtin;
 mod vpc_firewall;
@@ -25,3 +26,6 @@ mod vpc_routers;
 mod vpc_subnets;
 mod vpcs;
 mod zpools;
+
+// This module is used only for shared data, not test cases.
+mod endpoints;
