+++ nexus/tests/output/uncovered-authz-endpoints.txt
@@ -1,0 +1,22 @@
+API endpoints with no coverage in authz tests:
+instance_network_interfaces_delete_interface (delete "/organizations/{organization_name}/projects/{project_name}/instances/{instance_name}/network-interfaces/{interface_name}")
+project_snapshots_delete_snapshot        (delete "/organizations/{organization_name}/projects/{project_name}/snapshots/{snapshot_name}")
+hardware_racks_get                       (get    "/hardware/racks")
+hardware_racks_get_rack                  (get    "/hardware/racks/{rack_id}")
+hardware_sleds_get                       (get    "/hardware/sleds")
+hardware_sleds_get_sled                  (get    "/hardware/sleds/{sled_id}")
+instance_disks_get                       (get    "/organizations/{organization_name}/projects/{project_name}/instances/{instance_name}/disks")
+instance_network_interfaces_get          (get    "/organizations/{organization_name}/projects/{project_name}/instances/{instance_name}/network-interfaces")
+instance_network_interfaces_get_interface (get    "/organizations/{organization_name}/projects/{project_name}/instances/{instance_name}/network-interfaces/{interface_name}")
+project_snapshots_get                    (get    "/organizations/{organization_name}/projects/{project_name}/snapshots")
+project_snapshots_get_snapshot           (get    "/organizations/{organization_name}/projects/{project_name}/snapshots/{snapshot_name}")
+subnet_network_interfaces_get            (get    "/organizations/{organization_name}/projects/{project_name}/vpcs/{vpc_name}/subnets/{subnet_name}/network-interfaces")
+sagas_get                                (get    "/sagas")
+sagas_get_saga                           (get    "/sagas/{saga_id}")
+session_me                               (get    "/session/me")
+timeseries_schema_get                    (get    "/timeseries/schema")
+spoof_login                              (post   "/login")
+logout                                   (post   "/logout")
+instance_network_interfaces_post         (post   "/organizations/{organization_name}/projects/{project_name}/instances/{instance_name}/network-interfaces")
+project_snapshots_post                   (post   "/organizations/{organization_name}/projects/{project_name}/snapshots")
+updates_refresh                          (post   "/updates/refresh")
