+++ nexus/tests/output/unexpected-authz-endpoints.txt
@@ -1,0 +1,1 @@
+API endpoints tested by unauthorized.rs but not found in the OpenAPI spec:
